large-app/requirements.txt
flask
flask-bootstrap
flask-moment
flask-wtf
flask-sqlalchemy
flask-migrate
flask-mail
python-dotenv